Reverted Added Business Management specialisations
Specialisations cannot be derived from programme. Management student would get a “Sorry, your programme is not supported yet” when the system tries to create an account for them. To support specialisations, the core model of this project must be updated with an extra class eg Specialisation that is dependent of Curriculum and user settable in the GUI. tl;dr there is no way (using Detour) to derive the user’s specialisation.